+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Changed
 
+-   [core] simplify typescript type for tick label format in `AxisTicks`
 -   [core] adjusted the default theme to disable pointer events on decorative items, e.g. grid lines
 -   [band] updated `BandSurface` to respond to mouse events and use a
     customizable 'dataComponent'